工作台接口和excel表格导出

+++ sky-server/src/main/java/com/sky/mapper/UserMapper.java
@@ -44,5 +44,5 @@ public interface UserMapper {
      * @param mp
      * @return
      */
-    Double countByMap(Map mp);
+    Integer countByMap(Map mp);
 }
